fixed warnings
git-svn-id: https://svn.opendtect.org/od/trunk@27430 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiSeis/uisegyimpdlg.cc
@@ -196,7 +196,7 @@ bool doImp( const FilePath& fp )
     }
 
     BufferString fullmaskfnm( maskfp.fullPath() );
-    int lnmoffs = mCast( float, strstr( fullmaskfnm.buf(), "*" ) - 
+    int lnmoffs = mCast( int, strstr( fullmaskfnm.buf(), "*" ) - 
 	                                   fullmaskfnm.buf() );
     const int orglen = fullmaskfnm.size();
     bool nofails = true;
